Move SubQueriesProcessor to another namespace
