+++ tests/Unit/QueryLanguage/Pql/ParserTest.php
@@ -23,7 +23,7 @@ use Pimcore\Bundle\GenericDataIndexBundle\Model\QueryLanguage\ParseResultSubQuer
 use Pimcore\Bundle\GenericDataIndexBundle\QueryLanguage\Pql\Lexer;
 use Pimcore\Bundle\GenericDataIndexBundle\QueryLanguage\Pql\Parser;
 use Pimcore\Bundle\GenericDataIndexBundle\SearchIndexAdapter\OpenSearch\QueryLanguage\PqlAdapter;
-use Pimcore\Bundle\GenericDataIndexBundle\SearchIndexAdapter\OpenSearch\QueryLanguage\PqlAdapter\SubQueriesProcessorInterface;
+use Pimcore\Bundle\GenericDataIndexBundle\SearchIndexAdapter\OpenSearch\QueryLanguage\SubQueriesProcessorInterface;
 use Pimcore\Bundle\GenericDataIndexBundle\Service\ElementServiceInterface;
 use Pimcore\Bundle\GenericDataIndexBundle\Service\SearchIndex\IndexEntityService;
 use Pimcore\Bundle\GenericDataIndexBundle\Service\SearchIndex\SearchIndexConfigServiceInterface;
